FIX: Allow configuration of IPMI privilege level
This change updates the ipmitool invocation and its wrapper functions
to pass an IPMI privilege level value (default 'ADMINISTRATOR').

Without this value, ipmitool requests will fail if the ipmi_user
does not have ADMINISTRATOR privilege; this privilege is typically
not required for the requests being made by this library.

This change depends on change 1076.

signed-off-by: dave kormann <davek@research.att.com>
Change-Id: I7bdd4f64a4704765761b86cf1d207c2a7849fe46